- Rename codegen files to be consistent with their output - Add underscore to separate words in file names
