+++ bin/codegen.dart
@@ -1,10 +1,10 @@
 import 'dart:io';
 
-import 'codegencontext.dart' show context;
-import 'templates/platformconstants.dart.dart' as dart_template;
-import 'templates/platformconstants.h.dart' as objc_header_template;
-import 'templates/platformconstants.java.dart' as java_template;
-import 'templates/platformconstants.m.dart' as objc_impl_template;
+import 'codegen_context.dart' show context;
+import 'templates/PlatformConstants.h.dart' as objc_header_template;
+import 'templates/PlatformConstants.java.dart' as java_template;
+import 'templates/PlatformConstants.m.dart' as objc_impl_template;
+import 'templates/platform_constants.dart.dart' as dart_template;
 
 typedef Template = String Function(Map<String, dynamic> context);
 
@@ -12,7 +12,7 @@ const String projectRoot = '../';
 
 Map<Template, String> toGenerate = {
   // input template method vs output file path
-  dart_template.$: '${projectRoot}lib/src/generated/platformconstants.dart',
+  dart_template.$: '${projectRoot}lib/src/generated/platform_constants.dart',
   java_template.$:
       '${projectRoot}android/src/main/java/io/ably/flutter/plugin/generated/PlatformConstants.java',
   objc_header_template.$:
