+++ bin/templates/platform_constants.dart.dart
